CAMEL-21183: BrowseEndpoint should have limit/filter to optimize returned data from component implementations

+++ core/camel-api/src/main/java/org/apache/camel/spi/BrowsableEndpoint.java
@@ -32,12 +32,12 @@ public interface BrowsableEndpoint extends Endpoint {
     /**
      * Maximum number of messages to browse by default.
      */
-    long getBrowseLimit();
+    int getBrowseLimit();
 
     /**
      * Maximum number of messages to browse by default.
      */
-    void setBrowseLimit(long browseLimit);
+    void setBrowseLimit(int browseLimit);
 
     /**
      * Return the exchanges available on this endpoint
