+++ core/camel-management/src/main/java/org/apache/camel/management/mbean/ManagedBrowsableEndpoint.java
@@ -43,12 +43,12 @@ public class ManagedBrowsableEndpoint extends ManagedEndpoint implements Managed
     }
 
     @Override
-    public long getBrowseLimit() {
+    public int getBrowseLimit() {
         return getEndpoint().getBrowseLimit();
     }
 
     @Override
-    public long queueSize() {
+    public int queueSize() {
         return getEndpoint().getExchanges().size();
     }
